+++ tests/test_seuratpreparing.py
@@ -2,12 +2,11 @@ import pytest  # noqa: F401
 
 from .conftest import run_process
 
-pytest_order = 2
+pytestmark = pytest.mark.order(2)
 
 
 @pytest.mark.forked
 def test_seuratpreparing(request):
-
     outdir = run_process(
         "SeuratPreparing",
         "seuratpreparing.config.toml",
@@ -18,7 +17,6 @@ def test_seuratpreparing(request):
 
 @pytest.mark.forked
 def test_seuratpreparing_sct(request):
-
     outdir = run_process(
         "SeuratPreparing",
         "seuratpreparing_sct.config.toml",